feat: start move to typescript
