+++ src/commands/ping.ts
